requirements.txt
einops==0.4.1
matplotlib==3.4.3
numpy==1.20.3
opencv_python==4.6.0.66
pandas==1.3.4
patchify==0.2.3
Pillow==8.4.0
Pillow==10.0.0
py_fibonacci==0.5.2
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.7.1
timm==0.6.13
torch==1.12.1
torchvision==0.13.1
tqdm==4.62.3
